requirements.txt
boto3==1.24.51
pydantic==1.9.1
fastapi==0.75.2
uvicorn==0.17.6
starlette==0.19.1
python-dotenv==0.20.0